+++ tests/unit/runtime/test_pre_push_validation_fix_pass_ancestry_helpers.py
@@ -1259,414 +1259,6 @@ async def test_commit_range_touches_path_decodes_byte_diffs(
 
 
 @pytest.mark.unit
-def test_diff_hunk_near_anchor_related_accepts_guard_window_before_line() -> None:
-    # Pure insert several lines before the review anchor (not line / line-1),
-    # but only when the insert shares the review line's enclosing definition.
-    same_fn = (
-        "def reviewed():\n"
-        "    a = 1\n"
-        "    b = 2\n"
-        "    c = 3\n"
-        "    d = 4\n"
-        "    e = 5\n"
-        "    f = 6\n"
-        "    do_work()\n"
-    )
-    assert (
-        ancestry._diff_hunk_near_anchor_related("@@ -3,0 +4,2 @@\n", 8, file_text=same_fn) is True
-    )
-    # Modifications near the anchor are not proximity evidence (call-site link only).
-    assert (
-        ancestry._diff_hunk_near_anchor_related("@@ -4,1 +4,1 @@\n", 8, file_text=same_fn) is False
-    )
-
-
-@pytest.mark.unit
-def test_diff_hunk_near_anchor_related_rejects_other_enclosing_def() -> None:
-    """Unrelated insert in a neighboring function must not count as near-anchor evidence."""
-    text = (
-        "def other():\n"
-        "    x = 1\n"
-        "    y = 2\n"
-        "\n"
-        "def reviewed():\n"
-        "    a = 1\n"
-        "    b = 2\n"
-        "    do_work()\n"
-    )
-    # Pure insert after line 2 inside other() — within the 12-line window of do_work.
-    assert ancestry._diff_hunk_near_anchor_related("@@ -2,0 +3,1 @@\n", 8, file_text=text) is False
-    assert ancestry._diff_hunk_near_anchor_related("@@ -3,0 +4,2 @@\n", 8, file_text="") is False
-
-
-@pytest.mark.unit
-def test_diff_hunk_near_anchor_related_rejects_distant_and_after() -> None:
-    text = "def reviewed():\n    do_work()\n"
-    assert ancestry._diff_hunk_near_anchor_related("@@ -1,0 +2,1 @@\n", 30, file_text=text) is False
-    assert (
-        ancestry._diff_hunk_near_anchor_related("@@ -20,0 +21,1 @@\n", 8, file_text=text) is False
-    )
-    assert ancestry._diff_hunk_near_anchor_related("@@ -3,0 +4,2 @@\n", 0, file_text=text) is False
-
-
-@pytest.mark.unit
-def test_callee_names_from_anchor_line_extracts_calls_and_filters_keywords() -> None:
-    assert ancestry._callee_names_from_anchor_line("    return helper(x)") == frozenset({"helper"})
-    assert ancestry._callee_names_from_anchor_line("if ready(x) and helper():") == frozenset(
-        {"ready", "helper"}
-    )
-    assert ancestry._callee_names_from_anchor_line("    return self.helper()") == frozenset(
-        {"helper"}
-    )
-    assert ancestry._callee_names_from_anchor_line("    if (x):") == frozenset()
-    assert ancestry._callee_names_from_anchor_line("    return None") == frozenset()
-    assert ancestry._callee_names_from_anchor_line("") == frozenset()
-    assert ancestry._callee_names_from_anchor_line("def reviewed():") == frozenset()
-    assert ancestry._callee_names_from_anchor_line("def reviewed(): return helper()") == frozenset(
-        {"helper"}
-    )
-
-
-@pytest.mark.unit
-def test_callee_refs_capture_optional_qualifier() -> None:
-    assert ancestry._callee_refs_from_anchor_line("    return self.helper()") == frozenset(
-        {("self", "helper")}
-    )
-    assert ancestry._callee_refs_from_anchor_line("    return helper()") == frozenset(
-        {(None, "helper")}
-    )
-
-
-@pytest.mark.unit
-def test_callee_refs_ignore_calls_inside_comments_and_string_literals() -> None:
-    # Call-shaped text in comments/literals must not become FIXED callee evidence.
-    assert ancestry._callee_refs_from_anchor_line("    # TODO: helper()") == frozenset()
-    assert ancestry._callee_refs_from_anchor_line("    x = 1  # helper()") == frozenset()
-    assert ancestry._callee_refs_from_anchor_line("    // helper()") == frozenset()
-    assert ancestry._callee_refs_from_anchor_line('    message = "helper()"') == frozenset()
-    assert ancestry._callee_refs_from_anchor_line("    message = 'helper()'") == frozenset()
-    assert ancestry._callee_refs_from_anchor_line("    message = `helper()`") == frozenset()
-    # Real call kept; literal decoy ignored.
-    assert ancestry._callee_refs_from_anchor_line(
-        '    return real_call("helper()")  # other()'
-    ) == frozenset({(None, "real_call")})
-    # JS private-field call is code, not a Python EOL comment.
-    assert ancestry._callee_refs_from_anchor_line("    return this.#helper()") == frozenset(
-        {(None, "helper")}
-    )
-
-
-@pytest.mark.unit
-def test_diff_text_changes_definition_names_detects_def_forms() -> None:
-    diff = "@@ -1,1 +1,1 @@\n-def helper():\n+def helper():\n"
-    # Signature-adjacent body change still names the def on a +/- line when present.
-    body_only = "@@ -2,1 +2,1 @@\n-    return 1\n+    return 2\n"
-    assert ancestry._diff_text_changes_definition_names(diff, frozenset({"helper"})) is True
-    assert ancestry._diff_text_changes_definition_names(body_only, frozenset({"helper"})) is False
-    assert ancestry._diff_text_changes_definition_names(diff, frozenset({"other"})) is False
-    assert ancestry._diff_text_changes_definition_names("", frozenset({"helper"})) is False
-    arrow = "@@ -1,1 +1,1 @@\n-const helper = () => {\n+const helper = () => {\n"
-    assert ancestry._diff_text_changes_definition_names(arrow, frozenset({"helper"})) is True
-
-
-@pytest.mark.unit
-def test_enclosing_definition_name_finds_nearest_def_above() -> None:
-    text = "def helper():\n    return 1\n\ndef reviewed():\n    return helper()\n"
-    assert ancestry._enclosing_definition_name(text, 2) == "helper"
-    assert ancestry._enclosing_definition_name(text, 5) == "reviewed"
-    assert ancestry._enclosing_definition_name(text, 0) is None
-    assert ancestry._enclosing_definition_name("", 1) is None
-
-
-@pytest.mark.unit
-def test_enclosing_definition_name_recognizes_arrow_assignment() -> None:
-    """JS/TS ``const helper = () =>`` bodies must count as enclosing definitions."""
-    text = (
-        "const helper = () => {\n  return 1;\n};\n\nfunction reviewed() {\n  return helper();\n}\n"
-    )
-    assert ancestry._enclosing_definition_name(text, 2) == "helper"
-    assert ancestry._enclosing_definition_name(text, 6) == "reviewed"
-    assert ancestry._enclosing_definition_identity(text, 2) == ("helper", 1)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_includes_arrow_assignment_body() -> None:
-    text = (
-        "const helper = () => {\n  return 1;\n};\n\nfunction reviewed() {\n  return helper();\n}\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=6, qualifier=None, name="helper"
-    ) == (1, 4)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_prefers_in_scope_target() -> None:
-    text = (
-        "def helper():\n"
-        "    return 99\n"
-        "\n"
-        "class Foo:\n"
-        "    def helper(self):\n"
-        "        return 1\n"
-        "\n"
-        "    def reviewed(self):\n"
-        "        return self.helper()\n"
-    )
-    # Attribute call resolves to the method inside Foo, not the module helper.
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=9, qualifier="self", name="helper"
-    ) == (5, 7)
-    # Bare call below would resolve to nearest preceding module/class-visible def.
-    bare = "def helper():\n    return 1\n\ndef reviewed():\n    return helper()\n"
-    assert ancestry._resolve_callee_definition_span(
-        bare, call_line=5, qualifier=None, name="helper"
-    ) == (1, 3)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_self_method_declared_after_call() -> None:
-    """Same-class ``self.helper()`` must resolve even when ``helper`` is declared later."""
-    text = (
-        "class Foo:\n"
-        "    def reviewed(self):\n"
-        "        return self.helper()\n"
-        "\n"
-        "    def helper(self):\n"
-        "        return 1\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=3, qualifier="self", name="helper"
-    ) == (5, 6)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_self_skips_nested_class_method() -> None:
-    """``self.helper()`` must not bind to a same-named method on a nested class."""
-    only_nested = (
-        "class Outer:\n"
-        "    def reviewed(self):\n"
-        "        return self.helper()\n"
-        "\n"
-        "    class Inner:\n"
-        "        def helper(self):\n"
-        "            return 1\n"
-    )
-    assert (
-        ancestry._resolve_callee_definition_span(
-            only_nested, call_line=3, qualifier="self", name="helper"
-        )
-        is None
-    )
-    with_own = (
-        "class Outer:\n"
-        "    def reviewed(self):\n"
-        "        return self.helper()\n"
-        "\n"
-        "    def helper(self):\n"
-        "        return 2\n"
-        "\n"
-        "    class Inner:\n"
-        "        def helper(self):\n"
-        "            return 1\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        with_own, call_line=3, qualifier="self", name="helper"
-    ) == (5, 7)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_skips_class_method() -> None:
-    """Bare ``helper()`` must not bind to a nearer same-named class method."""
-    text = (
-        "def helper():\n"
-        "    return 99\n"
-        "\n"
-        "class Foo:\n"
-        "    def helper(self):\n"
-        "        return 1\n"
-        "\n"
-        "    def reviewed(self):\n"
-        "        return helper()\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=9, qualifier=None, name="helper"
-    ) == (1, 3)
-    only_method = (
-        "class Foo:\n"
-        "    def helper(self):\n"
-        "        return 1\n"
-        "\n"
-        "    def reviewed(self):\n"
-        "        return helper()\n"
-    )
-    assert (
-        ancestry._resolve_callee_definition_span(
-            only_method, call_line=6, qualifier=None, name="helper"
-        )
-        is None
-    )
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_later_toplevel_helper() -> None:
-    """Module-level helpers declared after the call site remain in scope for bare calls."""
-    text = "def reviewed():\n    return helper()\n\ndef helper():\n    return 1\n"
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=2, qualifier=None, name="helper"
-    ) == (4, 5)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_prefers_nested_helper() -> None:
-    """Nested helpers defined before the call beat same-named top-level defs."""
-    text = (
-        "def helper():\n"
-        "    return 99\n"
-        "\n"
-        "def reviewed():\n"
-        "    def helper():\n"
-        "        return 1\n"
-        "    return helper()\n"
-    )
-    # Nested helper ends at its lexical body, not the sibling ``return helper()``.
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=7, qualifier=None, name="helper"
-    ) == (5, 6)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_enclosing_function_helper() -> None:
-    """Bare calls must resolve helpers defined in an enclosing function scope."""
-    text = (
-        "def outer():\n"
-        "    def helper():\n"
-        "        return 1\n"
-        "    def reviewed():\n"
-        "        return helper()\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=5, qualifier=None, name="helper"
-    ) == (2, 3)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_prefers_inner_enclosing_helper() -> None:
-    """Innermost enclosing helper wins over an outer same-named helper."""
-    text = (
-        "def outer():\n"
-        "    def helper():\n"
-        "        return 99\n"
-        "    def mid():\n"
-        "        def helper():\n"
-        "            return 1\n"
-        "        def reviewed():\n"
-        "            return helper()\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=8, qualifier=None, name="helper"
-    ) == (5, 6)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_skips_sibling_nested_helper() -> None:
-    """Helpers local to a sibling nested def are not in scope for bare calls."""
-    text = (
-        "def outer():\n"
-        "    def sibling():\n"
-        "        def helper():\n"
-        "            return 1\n"
-        "    def reviewed():\n"
-        "        return helper()\n"
-        "\n"
-        "def helper():\n"
-        "    return 99\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=6, qualifier=None, name="helper"
-    ) == (8, 9)
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_bare_call_indented_module_helper() -> None:
-    """Module-scope helpers under ``if`` (indent > 0) remain bare-callable."""
-    text = (
-        "if True:\n"
-        "    def helper():\n"
-        "        return 1\n"
-        "    def reviewed():\n"
-        "        return helper()\n"
-    )
-    assert ancestry._resolve_callee_definition_span(
-        text, call_line=5, qualifier=None, name="helper"
-    ) == (2, 3)
-    if_body = "if True:\n    def helper():\n        return 1\n    x = helper()\n"
-    assert ancestry._resolve_callee_definition_span(
-        if_body, call_line=4, qualifier=None, name="helper"
-    ) == (2, 3)
-
-
-@pytest.mark.unit
-def test_iter_definition_spans_stops_at_module_level_assignment_dedent() -> None:
-    """Ordinary dedents (not only the next def) end a definition span."""
-    text = "def helper():\n    return 1\n\nX = 1\n\ndef reviewed():\n    return helper()\n"
-    spans = {
-        name: (start, end) for name, start, end, _indent in ancestry._iter_definition_spans(text)
-    }
-    assert spans["helper"] == (1, 3)
-    assert spans["reviewed"] == (6, 7)
-    helper_span = ancestry._resolve_callee_definition_span(
-        text, call_line=7, qualifier=None, name="helper"
-    )
-    assert helper_span == (1, 3)
-    # Unrelated module assignment must not count as FIXED callee-body evidence.
-    assert (
-        ancestry._diff_hunk_overlaps_line_span("@@ -4,1 +4,1 @@\n-X = 1\n+X = 2\n", *helper_span)
-        is False
-    )
-
-
-@pytest.mark.unit
-def test_resolve_callee_definition_span_rejects_unsupported_qualifier() -> None:
-    """Non-self/cls receivers must fail closed, not bind an unrelated bare def."""
-    text = "def send():\n    return 99\n\ndef reviewed():\n    return client.send()\n"
-    assert ancestry._callee_refs_from_anchor_line("    return client.send()") == frozenset(
-        {("client", "send")}
-    )
-    assert (
-        ancestry._resolve_callee_definition_span(text, call_line=5, qualifier="client", name="send")
-        is None
-    )
-
-
-@pytest.mark.unit
-def test_diff_hunk_overlaps_definition_span() -> None:
-    assert ancestry._diff_hunk_overlaps_line_span("@@ -2,1 +2,1 @@\n", 1, 3) is True
-    assert ancestry._diff_hunk_overlaps_line_span("@@ -10,1 +10,1 @@\n", 1, 3) is False
-    assert ancestry._diff_hunk_overlaps_line_span("@@ -2,0 +3,1 @@\n", 1, 3) is True
-
-
-@pytest.mark.unit
-def test_diff_hunk_related_line_evidence_combines_exact_and_near() -> None:
-    same_fn = (
-        "def reviewed():\n"
-        "    a = 1\n"
-        "    b = 2\n"
-        "    c = 3\n"
-        "    d = 4\n"
-        "    e = 5\n"
-        "    f = 6\n"
-        "    do_work()\n"
-    )
-    exact = "@@ -8,1 +8,1 @@\n-    do_work()\n+    do_work(1)\n"
-    near = "@@ -3,0 +4,2 @@\n+    if not ready:\n+        return\n"
-    distant = "@@ -40,1 +40,1 @@\n-    other()\n+    other(1)\n"
-    assert ancestry._diff_hunk_related_line_evidence(exact, 8, file_text=same_fn) is True
-    assert ancestry._diff_hunk_related_line_evidence(near, 8, file_text=same_fn) is True
-    assert ancestry._diff_hunk_related_line_evidence(distant, 8, file_text=same_fn) is False
-    assert ancestry._diff_hunk_related_line_evidence(near, 0, file_text=same_fn) is False
-
-
-@pytest.mark.unit
 async def test_commit_range_in_item_scope_accepts_blank_item_path() -> None:
     runner = SimpleNamespace(_deps=SimpleNamespace(runner=FakeCommandRunner()))
     assert (
